Refactor step 2
