+++ src/index.ts
@@ -1,70 +1,15 @@
-import { Template, getDefaultFont, PDFME_VERSION } from '@pdfme/common';
+import { Template } from '@pdfme/common';
 import { Designer, Form, Viewer } from '@pdfme/ui';
 import { generate } from '@pdfme/generator';
 
+import { defaultTemplate, themeSettings } from './constants';
+import { createTemplate, downloadJsonFile, getAvailableFonts, mapIssueDataToInput, mapIssueDataToTemplate, validateLocale } from './helper';
 import { getPlugins } from './schemas';
-import { validateLocale } from './types';
 
 import type { DesignerOptions, FormOptions, GeneratorOptions } from './types';
 
 declare const issueData: any;
-declare const embeddedFonts: any[];
 declare const pluginSettings: any;
-
-const themeSettings = {
-  token: {
-    colorPrimary: '#f1515c'
-  },
-};
-
-const defaultTemplate: Template = {
-  basePdf: { width: 210, height: 297, padding: [0, 0, 0, 0] },
-  schemas: [{} as Record<string, any>],
-  pdfmeVersion: PDFME_VERSION
-};
-
-/**
- * Checks if a value is valid.
- * @param value - The value to check.
- * @returns Whether the value is valid or not.
- */
-function isValidValue<T>(value: T | null | undefined | ''): value is T {
-  return value !== undefined && value !== null && value !== '';
-}
-
-/**
- * Get available fonts
- * @returns A promise that resolves to the available fonts.
- */
-async function getAvailableFonts(): Promise<Record<string, any>> {
-  const availableFonts = getDefaultFont();
-
-  for (const font of embeddedFonts) {
-    const response = await fetch(font.url);
-    const arrayBuffer = await response.arrayBuffer();
-
-    availableFonts[font.name] = {
-      data: arrayBuffer,
-      // include fallback and subset options if necessary
-    };
-  }
-
-  return availableFonts;
-}
-
-/**
- * Creates a template object based on the provided partial template.
- * @param template - The partial template object.
- * @returns The final template object.
- */
-function createTemplate(template: Partial<Template> = {}): Template {
-  const finalTemplate: Template = {
-    basePdf: isValidValue(template.basePdf) ? template.basePdf : defaultTemplate.basePdf,
-    schemas: isValidValue(template.schemas) ? template.schemas : defaultTemplate.schemas,
-    pdfmeVersion: isValidValue(template.pdfmeVersion) ? template.pdfmeVersion : defaultTemplate.pdfmeVersion
-  };
-  return finalTemplate;
-}
 
 /**
  * Opens the PDFMe Designer in the specified container.
@@ -119,14 +64,17 @@ export async function openViewer({
 
   const templateData = createTemplate(template) as Template | any;
   const pluginData = getPlugins({ fieldKeyOptions, fieldFormatOptions });
-  const inputData = mapIssueDataToTemplate(issueData, templateData) as Record<string, any>[]  || [{}];
+  const inputData = mapIssueDataToInput(issueData, templateData) as Record<string, any>[]  || [{}];
+
+  // Workaround for read-only mode, which does not accept inputs
+  const updatedTemplateData = mapIssueDataToTemplate(issueData, templateData) as Template | any;
 
   let instance;
 
   if (editing) {
     instance = new Form({
       domContainer: container,
-      template: templateData,
+      template: updatedTemplateData,
       plugins: pluginData,
       inputs: inputData,
       options: {
@@ -138,7 +86,7 @@ export async function openViewer({
   } else {
     instance = new Viewer({
       domContainer: container,
-      template: templateData,
+      template: updatedTemplateData,
       plugins: pluginData,
       inputs: inputData,
       options: {
@@ -152,25 +100,6 @@ export async function openViewer({
   // Return the Viewer or Form instance
   return instance;
 }
-
-/**
- * Downloads a JSON file containing the template data.
- * @param json - The template data to download.
- * @param title - The title of the file.
- */
-const downloadJsonFile = (json: unknown, title: string) => {
-  if (typeof window !== 'undefined') {
-    const blob = new Blob([JSON.stringify(json)], {
-      type: 'application/json',
-    });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = `${title}.json`;
-    link.click();
-    URL.revokeObjectURL(url);
-  }
-};
 
 /**
  * Downloads the current template as a JSON file.
@@ -250,56 +179,3 @@ export async function generatePdf({
 
   return blob;
 }
-
-/**
- * Retrieves the value from a nested object based on a dot-separated key path.
- * @param obj - The object to retrieve the value from.
- * @param keyPath - The dot-separated key path.
- * @returns The value at the specified key path, or undefined if not found.
- */
-function getNestedValue(obj: any, keyPath: string): any {
-  return keyPath.split('.').reduce((acc, key) => {
-    if (acc && acc[key] !== undefined) {
-      return acc[key];
-    } else {
-      console.warn(`Key not found: ${keyPath} at ${key}`);
-      return pluginSettings.default_placeholder_empty || '';
-    }
-  }, obj);
-}
-
-/**
- * Maps issue data to the template fields.
- * @param issueData - The issue data object.
- * @param template - The template object.
- * @returns The mapped inputs object.
- */
-function mapIssueDataToTemplate(issueData: any, template: Template) {
-  // console.log('Issue Data:', issueData);
-  // console.log('Template:', template);
-  const mappedInputs: Record<string, any>[] = [];
-
-  template.schemas.forEach((page: any, pageIndex: number) => {
-    // console.log('Page', pageIndex, page);
-    const mappedPage: Record<string, any> = {};
-
-    Object.keys(page).forEach((fieldKey: string) => {
-      // console.log('Field:', fieldKey, field);
-      const field = page[fieldKey];
-
-      if (field.field_key) {
-        const value = getNestedValue(issueData.issue, field.field_key);
-
-        if (value !== undefined) {
-          mappedPage[fieldKey] = value;
-        }
-      }
-    });
-
-    mappedInputs.push(mappedPage);
-  });
-
-  // Ensure the mappedInputs array contains only one object with all the fields
-  return [Object.assign({}, ...mappedInputs)];
-}
-
