+++ src/types.ts
@@ -1,5 +1,6 @@
 import { Template } from '@pdfme/common';
 import { Form, Viewer } from '@pdfme/ui';
+import { supportedLocales } from './constants';
 
 // Common Options Interfaces
 export interface FieldOption {
@@ -46,16 +47,5 @@ export interface GeneratorOptions extends CommonOptions {
 }
 
 // Supported Locales
-export const supportedLocales = ['en', 'ja', 'ar', 'th', 'it', 'pl', 'zh', 'ko', 'de', 'es', 'fr'] as const;
 export type SupportedLocale = typeof supportedLocales[number];
 
-/**
- * Validates the given locale and returns a supported locale.
- * If the given locale is not supported, 'en' (English) is returned as the default locale.
- *
- * @param locale - The locale to validate.
- * @returns The validated supported locale.
- */
-export function validateLocale(locale: string): SupportedLocale {
-  return supportedLocales.includes(locale as SupportedLocale) ? locale as SupportedLocale : 'en';
-}
